backend/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
opencv-python==4.8.0.74
rembg==2.0.38
Pillow==10.0.0
reportlab==4.0.4
numpy==1.24.3
python-dotenv==1.0.0